# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  payments.WebhookInput:
    properties:
      invoice_id:
        type: string
      status:
        type: string
      transaction_id:
        type: string
    type: object
  response.ErrorEnvelope:
    properties:
      error:
        properties:
          code:
            type: string
          details: {}
          message:
            type: string
        type: object
    type: object
...
info:
  contact: {}
  description: Backend API for PharmaLink B2B marketplace
  title: PharmaLink API
  version: "1.0"
paths:
...
  /orders/:
    get:
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: order payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create order (Pharmacy)
      tags:
      - orders
  /orders/{id}:
    get:
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get order by id
      tags:
      - orders
  /orders/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      - description: status payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - orders
  /payments/access:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get current access pass
      tags:
      - payments
  /payments/invoice:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create payment invoice
      tags:
      - payments
  /payments/webhook:
    post:
      consumes:
      - application/json
      parameters:
      - description: HMAC-SHA256 signature (hex)
        in: header
        name: X-Signature
        required: true
        type: string
      - description: webhook payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payments.WebhookInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: Payment webhook callback
      tags:
      - payments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"